Add files via upload
CRUD (Create, Read, Update, Delete) operations are fundamental actions in software development that deal with data manipulation in a persistent storage system, such as a database. In the context of ASP.NET MVC (Model-View-Controller), which is a web application framework developed by Microsoft, CRUD operations are typically associated with database interactions.